extract common factory functions
